Changes made by Alan: Modified the command line options in pymorph.py. Now the program will overwrite values in the config.py file if the options differ from the default. Otherwise, if options are default, only adds the value to config if it is not defined there.
Also added a vary_bulge command. This will give config.py control over setting 
bulge to pure DeVacouleurs (4.0) or allowing the index to vary (perform a Sersic
fit).

the vary_bulge parameter was added to the config.py.example in the Galfit Controls section.

Currently this control only seems to work if detail = 0